Tile Branch * Compile fixes for MSVC, patch by Andrew Wiggin (ender79). Thanks a lot! * Fixed scons compile on win, missing pthreads include.

+++ source/blender/blenlib/BLI_math_base.h
@@ -78,10 +78,10 @@
 #endif
 
 #ifndef sqrtf
-#define sqrtf(a) ((float)sqrt(a))
+#define sqrtf(a) ((float)sqrt((float)a))
 #endif
 #ifndef powf
-#define powf(a, b) ((float)pow(a, b))
+#define powf(a, b) ((float)pow((float)a, (float)b))
 #endif
 #ifndef cosf
 #define cosf(a) ((float)cos(a))
